move and rename patient contactInfo and address format functions
